Made layout of program; began on crossword generator
Defined the `Directions` class to specify the value of the "across" and "down" orientations of words in the crossword. The `Style` class specifies  what character represents an empty cell. The `Restrictions` class will eventually contain a regex that removes all non-language characters from a string to ensure the crossword contains no symbols, punctuation, whitespace, etc.

Additionally, 4 errors were defined as these will likely need to be handled in the future.

The `Crossword` class is still empty.