Added 7 methods to the Crossword class
Added the dunder `str` method to represent the class as a stringwhen printed

Added `generate` which is called on an instance of the calss to create and populate `self.grid`.

Added `_find_first_word_placement` to determine the middle indexes of `self.grid` for which to place the first word at.

Added `_find_insertion_coords` to determine all possible placements for a given word.

Added `_add_clue` and `_add_data` which append important information for all the words inserted to the crossword (orientation, intersections, position). `_add_clue` is friendly for the user to view and is displayed with by printing an instance of the Crossword class.

Added `_populate_grid` to find the best placement for a word based on intersections (using `_find_insertion_coords`). Words with no intersections are reinserted when `_populate_grid` is recursed. This function also appends placement information to self.clues and self.data (using `_add_clue` and `_add_data`.