Added intrinsic documentation and finalised the Crossword class
Added extensive intrinsic documentation including docstrings for `Crossword` and its methods.

Added `_prune_placements_for_readability` which eliminates word placements that are directly adjacent to another word. This change drastically improves the overall appearance of the crossword.

A minor change was made in `_populate_grid` to accompany for the new method that was introduced.