Added 3 static methods to the CrosswordHelper class
The `CrosswordHelper` class features 3 static methods that assist the `Crossword` class and the user

`find_best_crossword` generates a large sum of crosswords depending on the amount of words in the crossword (determined by loading `attempts.json` into a dictionary). This function does not require the user to call `generate()` and instead returns the crossword object directly.

`load_definitions` returns a specified crossword file JSON into a dictionary that is used by Crossword to insert words into `self.grid`.

`_load_attempts` returns the `attempts.json` file which relates the amount of words a crossword contains to the appropriate amount of iterations that `find_best_crossword` should run for before returning the best found crossword.

Driver code at the bottom of the script was added to showcase an example usage of `cword_gen.py`.