Added 104 languages to locales and updated main.py
Added the `Home` class that will contain the home screen widgets. Not much functionality is present.

The driver function in `main.py` reads the program's newly added `config.ini` which specifies various different attributes that are handled by the constructor of `Home`.

Also added `locale_utils.py` to parse `googletrans.LANGCODES` to be compatible for use with `pybabel init`.