Reorganised homescreen methods and added more global settings
Split the `_make_frames` and `_make_content` so they each have their own function to place themselves.

Added the layout for the `CrosswordBrowser` class which will let the user view all the available crosswords definition sets.

`change_appearance` and `change_scale` have been implemented to allow the user to switch appearance and increase or decrease scale with optionmenus.

Added some additional informational files for each crossword (all the files related to a crossword will now be stored in a folder with the crossword's name). This will help make the crosswords preview block in the crossword browser more informative.