Localised constants and errors and improved the Crossword Browser
Put all constants from across all the scripts into one file (`constants.py`) from where they can be imported as required. The same was done with the errors, which were put into `errors.py`. When `main.py` requires error detection, the errors will be added to `errors.py` as required.

Made `close_cword_browser` and the process of generating the content on the screen more modular by grouping the making and placing functions (for both frames and other widgets).

Created widgets for `CrosswordBrowser` and a function that instantiates HorizontalScrollFrame for n number of times (where n is the amount of available crossword files)

`HorizontalScrollFrame` creates a canvas with a horizontal scrollbar that controls frames within that canvas. Each frame contains an instance of `CrosswordInfoBlock`,  which displays the crosswords name, difficulty, and total word count (this data is read from `cword_folder/info.json`

`cfg` is no longer global, instead, it is passed to `Home` when it is instantiated. Child classes can access it when they inherit from `Home`.

`_determine_true_appearance` is a function that unfortunately had to be made because ctk.CTkCanvas cannot switch its foreground colour based on what the user sets with ctk.set_appearance_mode.

`_load_cword_info` retrieves the `info.json` file from a given crossword directory.