Annotated `main.py`, begin developing `cword_webapp`
`main.py` is now completely type annotated and intrinsically documented. The `HorizontalScrollFrame` class was removed and its functionality was replaced with a ctk.CTkScrollableFrame.

Added `requirements.txt`.

Began working on the interactive crossword webapp (using Flask and Jinja2) that will allow the user to fill in a generated crossword. Currently, only the html and css for viewing a grid of a variable side length has been written.