Finished html and css for crossword webapp
The Flask webapp can now run concurrently to the main tkinter application by using the multiprocessing module.

`render_template` contains a large amount of data from the crossword and other classes that will help the templated html and the javascript in accurately visualising the crossword and handling any interactions with it.

The css in `index.html` is at its final stages. It has been expanded to make the program aesthetically pleasing and professional.

Some extra bits of documentation were also added to various modules.