Developed `script.js`
`script.js` now contains functions that allow the user to interact with the program.

The event listener for `keydown` detects and validates proper user input, then places the entered character in currently focused cell, then shifts the input across or down.

`onDefinitionsListItemClick` allows the user to set their focus to the start of a word by clicking on its definition.

Currently, `checkIfCrosswordIsComplete` does not work. Another issue is the `keydown` listener removing all inner html of a cell when modifying its `innerText`, including the number label.

The html in `index.html` and some css within the same file was expanded upon to accompany for the new javascript functionality.