Simplified script.js and added backspace and click functionality
`checkIfCrosswordIsComplete` uses the `every` method to easily check if a crossword is complete.

The grid now responds to the users click and always ensures that only a single cell can be highlighted a different colour to show input focus (before, the old input focus did not revert its colour).

`onDefinitionsListItemClick` now sets the user's input focus to the beginning of the word that was clicked.

Regarding `index.html`, some changes were made to make it more readable and clean.

`babel.cfg` was made to specify what files to extract translatable strings from when using pybabel.