Updated colour schemes in the web app
Proper colour schemes were added to the web app based on the `Colours` class in `constants.py` (which received some minor changes in colour hex values). This will ensure that there is consistent colouring in both the customtkinter GUI and the web app. Additionally, a custom background from `https://css-pattern.com` was added.

Removed `Pillow==10.0.0` from `requirements.txt` as this will cause an error with the 10.2.0 version being in the file as well.