Minor bug fix in `script.js`
Prior to this bug fix in `script.js`, the input direction would automatically alternate when clicking at an intersection in the crossword grid. Now, you have to click a second time to alternate the direction.

Also added customtkinter button colours to Constants (and code to retrieve them in `_get_colour_palette_for_webapp` in `main.py`). When buttons are eventually added to the web app.