Moved css to `styles.css`
The css that was original in `index.html` has been linked and is now in its own file in the static directory of the web app. The root variables are still in `index.html` so that the new css file can access them.

Finally figured out how to properly parse a dictionary in JavaScript, so `colour_palette` is now a dictionary (in the html) and `json_colour_palette` is a JSON dumps object that is parsed by the JavaScript. Using a dictionary for the colour palette makes it much more obvious what colours the web app is using.