Added escape, space and arrow key functionality to the web app
The `keydown` event listener can now handle spacebar presses, as well as the escape key being pressed by removing the word and cell focus and reassigning `cellCoords` to `null`. Pressing the spacebar assigns a string with a single whitespace in it, which is accompanied for in the `isEmpty` function. It could be set to an empty string but this caused some problems when deleting.

`handleArrowPress` is a new function that allows the user to navigate the grid and switch directions using the arrow keys. It disregards any input that would shift the focus into a `.empty_cell` element.

`constants.py` now uses `os.path.join` for its `Paths` class instead of using hardcoded paths.

Renamed `data.json` to `definitions.json` for the capitals crossword as I forgot to do this earlier.