Added current word highlighting to `script.js`
The script now has the ability to highlight the definitions list item (or clue) corresponding to what word you are currently focused on. Some functions had to be added to achieve this, such as `updateCurrentWord` which retrieves the word the user is focused on. This word is used in the query selector for the `data-word` attribute (that is now provided in the HTML) in the `getDefinitionsListItemFromWord` function to find the correct clue to highlight.

Moved `handleSpacebarPress` and `handleEscapePress` into their own functions.

`handleArrowPress` was changed to only alternate the direction when pressing the opposite direction at an intersection instead of alternating and immediately shifting the cell, to keep the behaviour consistent with click interactions.

Some additional rules were added to `styles.css` to make the `.definitions` div evenly spaced and padded.