Added completion popup and tab support to the web app
An animated popup box will now slide into the centre the screen when completing a crossword. Some inline JS that toggles the the `active` classes of the popup elements was coded in `index.html`.

A bunch of CSS was added to achieve a stylish and clean popup design.

Additionally, tab support was added to the program. Using tab allows the user control what definitions/clue element they are focused on. Pressing enter while they are focused on it will emulate the function of a mouse, hence allowing the user to select words without needing a pointer. The popup close button is also immediately focused after the popup transitions into the page, allowing for fast keyboard user interaction.

Did some small additional type annotations to `main.py`.

Made some parts of `cword_gen.py` shorter. It is still a mess though.