Added translations catalogue
`translation_utils.py` was created to translate all untranslated `msgstr` entries in all the `locales` directory. The `google.cloud` module `translate_v2` was used to quickly and efficient create these translations.

Some changes were made to `app.py` to properly integrate it with the rest of the app using `flask-babel`.

All strings that are displayed to the user in `index.html` are now marked as gettext translation strings.