Update `base_crosswords` and some other modules
Forgot to push the `base_crosswords` directory which is required for the program to operate.

Some minor changes were made across the modules:
- Renamed some classes and attributes, as well as updated the variable names for the crossword images to better fit their shortened name

- Added some extra font classes to `main.py`

A major change was made to the display of crosswords in the crossword browser. Each category is now it's own block (instantiated from the new `CrosswordCategoryBlock` class) that manages its parent blocks (the crosswords). Therefore, some of the functionality from `CrosswordBrowser` has been transferred down to `CrosswordCategoryBlock`.