Added reveal, check and clear buttons to the web app
The web app now has 3 buttons (Reveal, Check and Clear) that open dropdowns, allowing the user to reveal/check the current cell, word or the grid, as well as clear the current word or the grid. `interaction.js` has been modified to properly serve this functionality to the user with two new functions (`handleCheckOrRevealButtonPress` and `handleClearButtonPress`).

An on-load popup was added to the web app to display the crossword name, category, word count and any failed insertions (with the adjusted word count).

The `getWordIndices` function in `interaction.js` is now a generator function.