Added keybinds to web app
`interaction.js` now has support for a few keybinds. Using combinations of "Shift" and "Enter", "Clear", or just "Enter" allows the user to reveal, check or clear a word respectively.

Broke down the function that reveals the cell/word/grid and isolated each main operation in its own function.

Slightly modified `styles.css`, mainly for the purpose of replacing fixed units with relative units.

The inline JavaScript in `index.html` now allows dropdowns to close if the user clicks anywhere but the dropdown. The HTML structure was also properly indented.