Added zooming functionality to the web app
zoomooz.js, a very useful jQuery plugin (check it out here: https://jaukia.github.io/zoomooz/) was used to implement zooming functionality in the web app.

The HTML structure of the program had to be modified to include `zoomViewport` and `zoomContainer` divs, as well as to include the script of zoomooz and jQuery.

A new checkbox was added to toggle the zoom.

To properly prevent zooming, a stupidly long amount of time was wasted trying multiple different implementations of this feature. Eventually, the `onCellClick` and `onDefinitionsListItemClick` were removed from the HTML command attribute, and instead were implemented via event listeners. This allowed for the use of the `event.stopPropagation` method, which prevents the jQuery listener from zooming if the click to zoom checkbox is disabled.

A bunch of new CSS and JS (to modify the crap CSS) was added to integrate the zooming feature as well as possible. Many issues were encountered, mainly overflow content due to the mess of new containers in the HTML.

To complement the new zooming feature, `styles.css` was modified to allow font scaling relative to that of the GUI.  For example, if you have a font scaling of 1.5 on the GUI, then all paragraphs in the web app, for example, will have 1.5x larger text.

The original inline JS in `index.html` was moved to `utils.js`. The only inline JS that will be written in `index.html` from now on will be related to scrolling, as it often requires jQuery, and I don't know how to make it work in its own JS file.

Finally, a favicon was also added (it is identical to the logo).