Update web app
Added the updateDefinitionsListPos function which triggers whenever a new word is selected. It ensures the definitions list is at the correct position so that the user can see their current word's definition. Also added a fallback value of 1 to max_attempts if the word count of the current crossword is not in attempts_db.