Merge most helper functions into utils.py
Added all of the CrosswordHelper functions into utils.py, as these functions are also used by main.py, so it makes sense for them to be in a central file. Almost all of the AppHelper functions were also added to this file, except for those that are related to i18n and are solely used within the main module.