Refactor: format whole codebase
Used the black formatter to format all python files. Additionally, some improvements were made to the code with the aims of improving conciseness and readability. The Prettier formatter was used to format all of the JavaScript, HTML and CSS code. Some extra documentation was added to interaction.js as well.