Improve crossword/crossword category viewing logic
Reduced the possibility of errors when viewing/loading crossword categories and crosswords. Additionally, some methods were added to utils.py to help the program infer the contents of an info.json file for a crossword directory with an empty info.json file, as well as randomly generate an info.json file for bare crossword category directories. This will allow the user more easily to create their own crosswords (if they really want to).