Add compound input feature to web app.
Added the ability for a user to type an unlimited amount of text in a cell by pressing Shift + 1 or using the new 'compound' button that was added. This allows users with keyboards such as kanji and hindi to combine different consonants/vowels together, as before, only a single character could be entered into a cell before the input focus was shifted. A user can confirm their input by pressing escape, enter or interacting with a dropdown. This will result in the character at index 0 of the compound input being placed into the cell.\n\nAlso slightly reorganised the placement of the buttons within the web app and added unicode arrows to the dropdowns to signify them being open or closed.\n\nAdded new translations for 'Compound'.\n\nIt is important that interaction.js as well as utils.js are thoroughly cleaned up in future commits as they are currently very unorganised and difficulty to follow.