Minor bug fixes to the web app
Removed the local jquery script as it caused bugs in production usage when testing. It will now be loaded from the internet as it was before.

Added the 'stopPropagation' method to the compound button onclick event, to prevent the event being picked up by the listener that removes the active compound input element when clicking anywhere other than a compound input element; as soon as the original onclick function created the compound input element, this other listener would remove it, thus making the button uselss.

Localised the 'border-right' attribute to only the 'definitions_a' div, not both, to prevent the appearance of two vertical dividers/borders.