Minor changes for structure and ergonomics of web app
Divided the HTML structure of the web app to be grouped by a 'left_side' and 'right_side' div (within the main container), improving the overall look of the web app.

Fixed some of the CSS after this change was made, primarily with the zoom containers. Also fixed the background colour of the compound input in light mode as previously it was difficulty to see.

Improved the zooming feature by making it follow you as you type into the grid or shift your focus with the arrow keys. Currently, if you quickly move in one direction (with arrow keys), then go back in the opposite direction, zoomooz will unzoom. I will try fixing this in the future if possible.

Added a incremental/decremental-based method to move through the definitions by clicking [Shift + ArrowUp] or [Shift + ArrowDown] to move to the next word and its definition.

Attempted to improve the zooming feature, as in some cases, the user must click twice to zoom. This issue was partially addressed with 'Interaction.handleDoubleClickForZoom', but it still is prevalent in some cases (not sure why).