Switch grid cookie to `localStorage` and improve `tabindex` feature
Storing the entire grid state and its classes was often too much data for a cookie to hold, so this data is now stored in the browser's local storage.

Reorganised the HTML so it follows a logical tab structure, as using a tabindex higher than 0 is not a good practice. The 'lang' attribute was also added to the '<html>' tag.

Allowed the user to tab to the grid and press enter to tab between all the non-empty cells.

Added some compatibility fixes to the CSS.