Add auto-language detection and system data storage
On the first launch, the program will attempt to initialise the GUI with the user's OS language.

Began storing user categories and config data in the OS document directory (I tried using the config/appdata dir, but it didn't work). If the documents cannot be accessed, the program will fallback to using package data.