Improve auto-word and compound input
Auto-word now shifts to the next word that has most likely been untouched by the user by checking if the first cell is filled or not. This method is only used when the user is shifting forward (not via Shift + ArrowDown). Otherwise, a straightforward shift is performed to the previous or next word.

Removing a compound input now calls 'crosswordCompletionHandler'.