Improve editor
Previously, if you selected a nondefault difficulty for a crossword, then defaulted the text in a form, the save button would be disabled (when it should remain enabled as the difficulty was changed). This was fixed.

Also added a try-except block for the 'edit' part of CrosswordPane._write, as renaming a crossword to a name that already exists would cause an error.