Improve naming of some GUI text
In the crossword browser, 'Load' was renamed to 'Create', 'Open' was renamed to 'Play', 'Terminate' was renamed to 'Destroy', and 'Make PDF' was renamed to 'Write to PDF'. These names are clearer and more descriptive. The coherence of a PDF-related messagebox string in utils.py was also improved.

The required translations for these changes were made and compiled, of course.